界面整修：自绘标题栏替换系统装饰
之前的窗口一直戴着系统标题栏。Windows 10 上它是一条灰底、硬直边、三个方形
按钮的系统栏，紧贴在一层 8pt 圆角卡片加浅灰底的面板上面——一个窗口两种设计
语言，接缝正好在眼睛最先落下的地方。

系统标题栏没法劝：DWMWA_CAPTION_COLOR 要 Windows 11 build 22000 起，
Win10 上根本没有颜色可设；按钮画在窗口的非客户区，应用拿不到那份画笔；按钮
形状由主题的视觉样式决定，只能换成系统的另一套。唯一出路是把整条换掉。

新增 src/ui/window_chrome.rs：

  * 标题栏（CAPTION_HEIGHT = 32，和 Windows 11 的标题栏同高）加三个窗口按钮
    （最小化 / 最大化-还原 / 关闭），用下面那套设计令牌画——同样的 space::M
    内边距、同样的 Palette::surface 底、同样的 control_hover 悬停色。
    关闭按钮用固定的 #C42B1C 而不是 palette.danger：后者是为「在白卡片上当
    文字」调的，暗色主题下偏亮，白字形压上去只有 2.8:1，而这个红有 5.4:1。

  * 无边框窗口没有缩放边框。winit 的 WM_NCCALCSIZE 把整个窗口当客户区，
    DefWindowProc 对每个像素都回 HTCLIENT，系统的边缘命中测试无从下手。
    补回来的是 8 个前层 Area 热区（四角加四边），按下发 BeginResize。
    做成 widget 而不是从 input 里判坐标是必须的：热区压在面板上，读 input
    会让一次按下既触发缩放、又漏给下面的按钮，而 BeginResize 之后鼠标已经
    交给 OS 的模态循环，按钮永远收不到松开、会卡在按下态。做成 widget 则由
    egui 按「最上层命中者胜」处理，和 Windows 的做法一致。

  * 撤掉装饰之后窗口既没有投影也没有边缘，补一圈 1pt 描边。颜色用
    border_strong——这是窗口和「桌面」之间的边，不是卡片和窗口之间的边，
    给卡片用的那道淡边贴到壁纸上就没了。

window_shape.rs：最大化时 SetWindowRgn(hwnd, None, true) 清掉圆角区域。
最大化窗口的四角就是屏幕的四个角，切圆角等于在桌面上啃四口。

panel.rs：MIN_PANEL_HEIGHT 480 提到 512（标题栏占的那 32 得补上），标题栏的
高度也进入 fit_to_content 的 chrome 算术。

顺手清掉此前遗留的 8 条 clippy 告警（device.rs / engine.rs / preset.rs，都是
机械的 as 冗余转换、meter_tick % 8 == 0、MUSIC.to_vec()），clippy 现在零告警。

测试：cargo test --lib --locked 58 passed / 1 ignored。新增 5 个窗口装饰测试，
其中「点关闭真的发 Close」那个是离线跑真帧——三帧 ctx.run_ui，按下与松开分在
两帧（点击本来就是这么来的），断言 ViewportCommand::Close 真的从 viewport
输出里出来，而不是只断言某个 Response 说自己被点了。

+++ src/device.rs
@@ -131,14 +131,14 @@ impl std::fmt::Display for DeviceState {
 
 impl From<DEVICE_STATE> for DeviceState {
     fn from(state: DEVICE_STATE) -> Self {
-        let raw = state.0 as u32;
-        if raw & DEVICE_STATE_ACTIVE.0 as u32 != 0 {
+        let raw = state.0;
+        if raw & DEVICE_STATE_ACTIVE.0 != 0 {
             DeviceState::Active
-        } else if raw & DEVICE_STATE_DISABLED.0 as u32 != 0 {
+        } else if raw & DEVICE_STATE_DISABLED.0 != 0 {
             DeviceState::Disabled
-        } else if raw & DEVICE_STATE_NOTPRESENT.0 as u32 != 0 {
+        } else if raw & DEVICE_STATE_NOTPRESENT.0 != 0 {
             DeviceState::NotPresent
-        } else if raw & DEVICE_STATE_UNPLUGGED.0 as u32 != 0 {
+        } else if raw & DEVICE_STATE_UNPLUGGED.0 != 0 {
             DeviceState::Unplugged
         } else {
             DeviceState::Unknown
